requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
pydantic==2.6.4
Jinja2==3.1.4
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiosqlite==0.19.0
redis==5.0.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
pytest==8.2.0
pytest-asyncio==0.23.5.post1
